# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a megrez backend server.",
        "title": "Megrez backend API",
        "termsOfService": "http://megrez.run",
        "contact": {
            "name": "Megrez",
            "url": "http://megrez.run",
            "email": "alkaidchen@qq.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/admin/attachments": {
            "get": {
                "description": "list attachments",
                "consumes": [
                    "application/json"
                ],
                "summary": "list attachments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "page num",
                        "name": "pageNum",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page size",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/errmsg.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/dto.Pagination"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "list": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/dto.AttachmentDTO"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/api/admin/upload": {
            "post": {
                "description": "upload attachment",
                "consumes": [
                    "application/json",
                    " text/plain",
                    " */*"
                ],
                "summary": "upload attachment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "type": "file",
                        "description": "file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/errmsg.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.Attachment"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dto.AttachmentDTO": {
            "type": "object",
            "properties": {
                "ext": {
                    "type": "string"
                },
                "fileName": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "size": {
                    "type": "string"
                },
                "thumbURL": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                },
                "uploadTime": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                }
            }
        },
...
        "dto.Pagination": {
            "type": "object",
            "properties": {
                "current": {
                    "type": "integer"
                },
                "list": {},
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "errmsg.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "msg": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "model.Attachment": {
            "type": "object",
            "properties": {
                "ThumbURL": {
                    "type": "string"
                },
                "ext": {
                    "type": "string"
                },
                "fileName": {
                    "type": "string"
                },
                "height": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "type": {
                    "type": "integer"
                },
                "uploadTime": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "width": {
                    "type": "integer"
                }
            }
        },
...
    }
}